Refactor asset/scene loading to use streams, update APIs
Refactored asset and scene loading to operate on streams instead of raw memory blocks, improving flexibility and efficiency. Replaced World.Clear(TimeData) with World.Reset() and updated all usages. Updated SceneManager and asset entry APIs to use streams, added new StreamUtility methods, and removed IDisposable from Scene. Refactored SystemManager to remove TimeData from lifecycle methods. Performed related code cleanups and updated tests.

+++ src/Runtime/Ghost.Core/Utilities/MemoryManager.cs
@@ -43,7 +43,7 @@ public unsafe class NativeMemoryManager<T> : MemoryManager<T>
         return new NativeMemoryManager<T>((T*)collection.GetUnsafePtr(), length);
     }
 
-    public static NativeMemoryManager<T> FromMemoryBlock(MemoryBlock memoryBlock, int start, int length)
+    public static NativeMemoryManager<T> FromMemoryBlock(MemoryBlock memoryBlock, nuint start, int length)
     {
         return new NativeMemoryManager<T>((T*)memoryBlock.GetUnsafePtr() + start, length);
     }
